requirements.txt
Flask-Assets
Flask-Redis
Flask-Script
Flask-SocketIO
gunicorn
